server/controller/src/test/java/ai/starwhale/mlops/api/LogControllerTest.java
/*
 * Copyright 2022 Starwhale, Inc. All Rights Reserved.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package ai.starwhale.mlops.api;

import static org.hamcrest.MatcherAssert.assertThat;
import static org.hamcrest.Matchers.allOf;
import static org.hamcrest.Matchers.is;
import static org.hamcrest.Matchers.isA;
import static org.hamcrest.Matchers.iterableWithSize;
import static org.hamcrest.Matchers.notNullValue;
import static org.mockito.ArgumentMatchers.anyLong;
import static org.mockito.ArgumentMatchers.anyString;
import static org.mockito.BDDMockito.given;
import static org.mockito.Mockito.mock;

import ai.starwhale.mlops.domain.task.TaskService;
import java.util.List;
import java.util.Objects;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.springframework.http.HttpStatus;

public class LogControllerTest {

    private LogController logController;

    private TaskService taskService;

    @BeforeEach
    public void setUp() {
        taskService = mock(TaskService.class);
        logController = new LogController(taskService);
    }

    @Test
    public void testOfflineLogs() {
        given(taskService.offLineLogFiles(anyLong()))
                .willReturn(List.of("offline", "log"));
        var resp = logController.offlineLogs(1L);
        assertThat(resp.getStatusCode(), is(HttpStatus.OK));
        assertThat(Objects.requireNonNull(resp.getBody()).getData(), allOf(
                notNullValue(),
                isA(List.class),
                is(iterableWithSize(2))
        ));
    }

    @Test
    public void testLogContent() {
        given(taskService.logContent(anyLong(), anyString()))
                .willReturn("content");
        var resp = logController.logContent(1L, "file");
        assertThat(resp.getStatusCode(), is(HttpStatus.OK));
        assertThat(resp.getBody(), is("content"));
    }
}
